Replaced obsolete FIBF_ARCHIVE and FIBF_SCRIPT flags with new flags.
Added actual "replace" functionality to COPY and MOVE functions when
a destination directory already exists. Directories were previously
merged when "Replace" was selected in a requester.

Fixed RENAME function to show new filename in a lister when a name
already exists and the name is changed for the file that is being
renamed.