Replaced obsolete FIBF_ARCHIVE and FIBF_SCRIPT flags with new flags.
Added actual "replace" functionality to COPY and MOVE functions when
a destination directory already exists and Tooltype REPLACEDIRS is set.

Fixed RENAME function to show new filename in a lister when a name
already exists and the name is changed for the file that is being renamed.

Fixed infinite recurse that occurred when user copied or moved a
directory into one of it's subdirectories.

Fixed skipall option broken by earlier changes.