Fixed directory sizes displayed in a lister when searching for a filename (or pattern) in multiple directories with the hunt command.
Fixed a hunt command failure to highlight the found file when the file
was located in a subdirectory of the directory being searched. When the
user selected the OK button in the "Shall I go there?" requester,
the hunt command would display a requester stating that the file was
not found but then load the directory where the file was located.